+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Abstract Factory pattern
+- Prototype pattern
 
 
 ## [2022-06-26]